+++ Data Admin/CHANGEDB.php
@@ -136,3 +136,9 @@ $count++;
 $sql[$count][0]="1.6.00" ;
 $sql[$count][1]="";
 
+//v1.6.01
+$count++;
+$sql[$count][0]="1.6.01" ;
+$sql[$count][1]="
+UPDATE `gibbonAction` SET `URLList`='records_manage.php,records_orphaned.php,records_duplicates.php,export_run.php' WHERE name='Manage Records' AND gibbonModuleID=(SELECT gibbonModuleID FROM gibbonModule WHERE name='Data Admin');end";
+